now working :)

+++ src/main/kotlin/com/voongc/data/Event.kt
@@ -2,7 +2,7 @@ package com.voongc.data
 
 open class Catalogue
 
-data class Event(val id:String,val category:String,val subCategory:String,val name:String,val startTime:Int, val displayed:Boolean, val suspended:Boolean): Catalogue()
+data class Event(val id:String,val category:String,val subCategory:String,val name:String,val startTime:Long, val displayed:Boolean, val suspended:Boolean): Catalogue()
 
 data class Market(val eventId: String,val id:String,val name:String,val displayed:Boolean,val suspended: Boolean) : Catalogue ()
